+++ client/src/components/cartProduct/cartProduct.tsx
@@ -7,6 +7,7 @@ import {
 } from '../../app/actions/cartActions/index';
 import {userInfo, userToken} from '../../app/reducers/registerReducer';
 import './cartProduct.css';
+import swal from 'sweetalert2';
 
 const CartProduct = (props) => {
   // eslint-disable-next-line react/prop-types
@@ -17,11 +18,25 @@ const CartProduct = (props) => {
   const token = useAppSelector(userToken);
   const dispatch = useAppDispatch();
   const handleDeleteProduct = () => {
-    dispatch(deleteProductFromCart(id, idOrder, userId, token));
-    alert('Product deleted successfully');
+    swal.fire({
+      title: 'Delete Product?',
+      text: 'Are you sure you want to delete this product from cart?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteProductFromCart(id, idOrder, userId, token));
+        swal.fire({
+          text: 'Product deleted!',
+          icon: 'info',
+        });
+      }
+    }).catch((err) => console.error(err));
   };
   useEffect(() => {
-    console.log('algo');
     const totalPrice = price * productAmount;
     dispatch(changeProductQuantity(userId, id, productAmount,
         totalPrice, stock));
